+++ crates/sncast/src/starknet_commands/account/mod.rs
@@ -192,12 +192,12 @@ fn generate_add_profile_message(
     keystore: Option<Utf8PathBuf>,
 ) -> Result<Option<String>> {
     if let Some(profile_name) = profile_name {
-        let url = rpc
-            .url
-            .clone()
-            .expect("the argument '--network' should not be used with '--add-profile' argument");
+        let url = &rpc.url.clone().unwrap_or_default();
+        let network = rpc.network;
+
         let config = CastConfig {
-            url,
+            url: url.clone(),
+            network,
             account: account_name.into(),
             accounts_file: accounts_file.into(),
             keystore,