+++ crates/sncast/tests/e2e/account/import.rs
@@ -275,7 +275,7 @@ pub async fn test_nonexistent_account_address() {
 #[test_case("--url", URL; "with_url")]
 #[test_case("--network", "sepolia"; "with_network")]
 pub async fn test_happy_case_add_profile(flag: &str, value: &str) {
-    let tempdir = tempdir().expect("Failed to create a temporary directory");
+    let tempdir = copy_config_to_tempdir("tests/data/files/correct_snfoundry.toml", None);
     let accounts_file = "accounts.json";
 
     let args = vec![
@@ -336,51 +336,16 @@ pub async fn test_happy_case_add_profile(flag: &str, value: &str) {
 
     let contents = fs::read_to_string(current_dir_utf8.join("snfoundry.toml"))
         .expect("Unable to read snfoundry.toml");
-    assert!(contents.contains("[sncast.my_account_import]"));
-    assert!(contents.contains("account = \"my_account_import\""));
-    assert!(contents.contains(&format!("url = \"{URL}\"")));
-    assert!(contents.contains(&format!(
-        "{} = \"{}\"",
-        flag.trim_start_matches("--"),
-        value
-    )));
-}
 
-#[tokio::test]
-pub async fn test_add_profile_with_network_arg() {
-    let tempdir = tempdir().expect("Failed to create a temporary directory");
-    let accounts_file = "accounts.json";
-
-    let args = vec![
-        "--accounts-file",
-        accounts_file,
-        "account",
-        "import",
-        "--network",
-        "sepolia",
-        "--name",
-        "my_account_import",
-        "--address",
-        "0x1",
-        "--private-key",
-        "0x2",
-        "--class-hash",
-        DEVNET_OZ_CLASS_HASH_CAIRO_0,
-        "--type",
-        "oz",
-        "--add-profile",
-        "my_account_import",
+    let expected_lines = [
+        "[sncast.my_account_import]",
+        "account = \"my_account_import\"",
+        "accounts-file = \"accounts.json\"",
+        &format!("{} = \"{}\"", flag.trim_start_matches("--"), value),
     ];
+    let expected_block = expected_lines.join("\n");
 
-    let snapbox = runner(&args).current_dir(tempdir.path());
-    let output = snapbox.assert();
-
-    assert_stderr_contains(
-        output,
-        indoc! {r"
-        error: the argument '--network <NETWORK>' cannot be used with '--add-profile <ADD_PROFILE>'
-    "},
-    );
+    assert!(contents.contains(&expected_block));
 }
 
 #[tokio::test]